+++ src/oid4vci/handlers/authorize.rs
@@ -72,7 +72,7 @@ use tracing::instrument;
 
 use crate::core::generate;
 use crate::oid4vci::endpoint::Request;
-use crate::oid4vci::provider::{Metadata, Provider, Subject};
+use crate::oid4vci::provider::{Metadata, Provider, StateStore, Subject};
 use crate::oid4vci::state::{Authorization, AuthorizedItem, Expire, ItemType, Stage, State};
 use crate::oid4vci::types::{
     AuthorizationDetail, AuthorizationDetailType, AuthorizationRequest, AuthorizationResponse,
@@ -80,7 +80,6 @@ use crate::oid4vci::types::{
 };
 use crate::oid4vci::{Error, Result};
 use crate::openid::oauth::GrantType;
-use crate::openid::provider::StateStore;
 
 /// Authorization request handler.
 ///
