+++ src/oid4vci/handlers/create_offer.rs
@@ -72,7 +72,7 @@ use tracing::instrument;
 
 use crate::core::generate;
 use crate::oid4vci::endpoint::Request;
-use crate::oid4vci::provider::{Metadata, Provider, Subject};
+use crate::oid4vci::provider::{Metadata, Provider, StateStore, Subject};
 use crate::oid4vci::state::{AuthorizedItem, Expire, ItemType, Offer, Stage, State};
 use crate::oid4vci::types::{
     AuthorizationCodeGrant, AuthorizationDetail, AuthorizationDetailType, CreateOfferRequest,
@@ -81,7 +81,6 @@ use crate::oid4vci::types::{
 };
 use crate::oid4vci::{Error, Result};
 use crate::openid::oauth::GrantType;
-use crate::openid::provider::StateStore;
 
 /// Invoke request handler generates and returns a Credential Offer.
 ///
