+++ src/oid4vci/handlers/deferred.rs
@@ -12,13 +12,12 @@ use tracing::instrument;
 
 use crate::oid4vci::endpoint::Request;
 use crate::oid4vci::handlers::credential::credential;
-use crate::oid4vci::provider::Provider;
+use crate::oid4vci::provider::{Provider, StateStore};
 use crate::oid4vci::state::{Stage, State};
 use crate::oid4vci::types::{
     CredentialResponseType, DeferredCredentialRequest, DeferredCredentialResponse,
 };
 use crate::oid4vci::{Error, Result};
-use crate::openid::provider::StateStore;
 
 /// Deferred credential request handler.
 ///
