+++ src/oid4vci/handlers/notification.rs
@@ -22,11 +22,10 @@
 use tracing::instrument;
 
 use crate::oid4vci::endpoint::Request;
-use crate::oid4vci::provider::Provider;
+use crate::oid4vci::provider::{Provider, StateStore};
 use crate::oid4vci::state::{Stage, State};
 use crate::oid4vci::types::{NotificationRequest, NotificationResponse};
 use crate::oid4vci::{Error, Result};
-use crate::openid::provider::StateStore;
 
 /// Notification request handler.
 ///
