+++ src/oid4vci/handlers/par.rs
@@ -12,11 +12,10 @@ use tracing::instrument;
 use crate::core::generate;
 use crate::oid4vci::endpoint::Request;
 use crate::oid4vci::handlers::authorize;
-use crate::oid4vci::provider::{Metadata, Provider};
+use crate::oid4vci::provider::{Metadata, Provider, StateStore};
 use crate::oid4vci::state::{PushedAuthorization, Stage, State};
 use crate::oid4vci::types::{PushedAuthorizationRequest, PushedAuthorizationResponse};
 use crate::oid4vci::{Error, Result};
-use crate::openid::provider::StateStore;
 
 /// Endpoint for the Wallet to push an Authorization Request when using Pushed
 /// Authorization Requests.
