+++ src/oid4vci/handlers/register.rs
@@ -3,11 +3,10 @@
 use tracing::instrument;
 
 use crate::oid4vci::endpoint::Request;
-use crate::oid4vci::provider::Provider;
+use crate::oid4vci::provider::{Provider, StateStore};
 use crate::oid4vci::state::State;
 use crate::oid4vci::types::{RegistrationRequest, RegistrationResponse};
 use crate::oid4vci::{Error, Result};
-use crate::openid::provider::StateStore;
 
 /// Registration request handler.
 ///
