+++ src/oid4vci/provider.rs
@@ -2,11 +2,14 @@
 
 use std::future::Future;
 
+use anyhow::Result;
+use chrono::{DateTime, Utc};
 use credibil_did::DidResolver;
 use credibil_infosec::Signer;
+use serde::Serialize;
+use serde::de::DeserializeOwned;
 
 use crate::oid4vci::types::{Client, Dataset, Issuer, Server};
-use crate::openid::provider::{self, StateStore};
 use crate::status::issuer::Status;
 
 /// Issuer Provider trait.
@@ -19,19 +22,19 @@ pub trait Provider:
 /// and `Server` metadata to the library.
 pub trait Metadata: Send + Sync {
     /// Client (wallet) metadata for the specified issuance client.
-    fn client(&self, client_id: &str) -> impl Future<Output = provider::Result<Client>> + Send;
+    fn client(&self, client_id: &str) -> impl Future<Output = Result<Client>> + Send;
 
     /// Credential Issuer metadata for the specified issuer.
-    fn issuer(&self, issuer_id: &str) -> impl Future<Output = provider::Result<Issuer>> + Send;
+    fn issuer(&self, issuer_id: &str) -> impl Future<Output = Result<Issuer>> + Send;
 
     /// Authorization Server metadata for the specified issuer/server.
     fn server(
         &self, server_id: &str, issuer_id: Option<&str>,
-    ) -> impl Future<Output = provider::Result<Server>> + Send;
+    ) -> impl Future<Output = Result<Server>> + Send;
 
     /// Used to dynamically register OAuth 2.0 clients with the authorization
     /// server.
-    fn register(&self, client: &Client) -> impl Future<Output = provider::Result<Client>> + Send;
+    fn register(&self, client: &Client) -> impl Future<Output = Result<Client>> + Send;
 }
 
 /// The Subject trait specifies how the library expects issuance subject (user)
@@ -43,11 +46,26 @@ pub trait Subject: Send + Sync {
     /// request.
     fn authorize(
         &self, subject_id: &str, credential_configuration_id: &str,
-    ) -> impl Future<Output = provider::Result<Vec<String>>> + Send;
+    ) -> impl Future<Output = Result<Vec<String>>> + Send;
 
     /// Returns a populated `Dataset` object for the given subject (holder) and
     /// credential definition.
     fn dataset(
         &self, subject_id: &str, credential_identifier: &str,
-    ) -> impl Future<Output = provider::Result<Dataset>> + Send;
+    ) -> impl Future<Output = Result<Dataset>> + Send;
+}
+
+/// `StateStore` is used to store and retrieve server state between requests.
+pub trait StateStore: Send + Sync {
+    /// Store state using the provided key. The expiry parameter indicates
+    /// when data can be expunged from the state store.
+    fn put(
+        &self, key: &str, state: impl Serialize + Send, expiry: DateTime<Utc>,
+    ) -> impl Future<Output = Result<()>> + Send;
+
+    /// Retrieve data using the provided key.
+    fn get<T: DeserializeOwned>(&self, key: &str) -> impl Future<Output = Result<T>> + Send;
+
+    /// Remove data using the key provided.
+    fn purge(&self, key: &str) -> impl Future<Output = Result<()>> + Send;
 }
