+++ src/oid4vp/types/request.rs
@@ -1,42 +1,18 @@
-//! # `OpenID` for Verifiable Presentations (`OpenID4VP`)
-
-use std::collections::HashMap;
 use std::fmt;
 use std::fmt::Debug;
-use std::future::Future;
 use std::io::Cursor;
 
-use anyhow::anyhow;
+use anyhow::{Result, anyhow};
 use base64ct::{Base64, Encoding};
-use credibil_did::DidResolver;
 pub use credibil_infosec::Signer;
 use qrcode::QrCode;
 use serde::de::{self, Deserializer, Visitor};
 use serde::ser::{SerializeMap, Serializer};
 use serde::{Deserialize, Serialize};
 
-use super::provider::{Result, StateStore};
 use crate::core::{Kind, urlencode};
-use crate::dif_exch::{InputDescriptor, PresentationDefinition, PresentationSubmission};
-use crate::openid::oauth::{OAuthClient, OAuthServer};
-use crate::w3c_vc::model::VerifiablePresentation;
-
-/// Verifier Provider trait.
-pub trait Provider: Metadata + StateStore + Signer + DidResolver + Clone {}
-
-/// The `Metadata` trait is used by implementers to provide `Verifier` (client)
-/// metadata to the library.
-pub trait Metadata: Send + Sync {
-    /// Verifier (Client) metadata for the specified verifier.
-    fn verifier(&self, verifier_id: &str) -> impl Future<Output = Result<Verifier>> + Send;
-
-    /// Wallet (Authorization Server) metadata.
-    fn wallet(&self, wallet_id: &str) -> impl Future<Output = Result<Wallet>> + Send;
-
-    /// Used by OAuth 2.0 clients to dynamically register with the authorization
-    /// server.
-    fn register(&self, verifier: &Verifier) -> impl Future<Output = Result<Verifier>> + Send;
-}
+use crate::dif_exch::{InputDescriptor, PresentationDefinition};
+use crate::oid4vp::types::Verifier;
 
 /// The Request Object Request is created by the Verifier to generate an
 /// Authorization Request Object.
@@ -539,248 +515,5 @@ impl<'de> Deserialize<'de> for RequestObjectResponse {
         }
 
         deserializer.deserialize_any(VisitorImpl)
-    }
-}
-
-/// Authorization Response request object is used by Wallets to send a VP Token
-/// and Presentation Submission to the Verifier who initiated the verification.
-#[derive(Clone, Debug, Default, Deserialize, Serialize, PartialEq, Eq)]
-pub struct ResponseRequest {
-    /// One or more Verifiable Presentations represented as base64url encoded
-    /// strings and/or JSON objects. The VP format determines the encoding.
-    /// The encoding follows the same format-based rules as for Credential
-    /// issuance (Appendix E of the [OpenID4VCI] specification).
-    ///
-    /// When a single Verifiable Presentation is returned, array syntax MUST NOT
-    /// be used.
-    ///
-    /// [OpenID4VCI]: (https://openid.net/specs/openid-4-verifiable-credential-issuance-1_0.html)
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub vp_token: Option<Vec<Kind<VerifiablePresentation>>>,
-
-    /// The `presentation_submission` element as defined in
-    /// [DIF.PresentationExchange]. It contains mappings between the
-    /// requested Verifiable Credentials and where to find them within the
-    /// returned VP Token.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub presentation_submission: Option<PresentationSubmission>,
-
-    /// The client state value from the Authorization Request.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub state: Option<String>,
-}
-
-impl ResponseRequest {
-    /// Create a `HashMap` representation of the `ResponseRequest` suitable for
-    /// use in an HTML form post.
-    ///
-    /// # Errors
-    /// Will return an error if any nested objects cannot be serialized and
-    /// URL-encoded.
-    pub fn form_encode(&self) -> anyhow::Result<HashMap<String, String>> {
-        let mut map = HashMap::new();
-        if let Some(vp_token) = &self.vp_token {
-            let as_json = serde_json::to_string(vp_token)?;
-            map.insert("vp_token".into(), urlencoding::encode(&as_json).to_string());
-        }
-        if let Some(presentation_submission) = &self.presentation_submission {
-            let as_json = serde_json::to_string(presentation_submission)?;
-            map.insert("presentation_submission".into(), urlencoding::encode(&as_json).to_string());
-        }
-        if let Some(state) = &self.state {
-            map.insert("state".into(), state.into());
-        }
-        Ok(map)
-    }
-
-    /// Create a `ResponseRequest` from a `HashMap` representation.
-    ///
-    /// Suitable for
-    /// use in a verifier's response endpoint that receives a form post before
-    /// passing the `ResponseRequest` to the `response` handler.
-    ///
-    /// # Errors
-    /// Will return an error if any nested objects cannot be deserialized from
-    /// URL-encoded JSON strings.
-    pub fn form_decode(map: &HashMap<String, String>) -> anyhow::Result<Self> {
-        let mut req = Self::default();
-        if let Some(vp_token) = map.get("vp_token") {
-            let decoded = urlencoding::decode(vp_token)?;
-            let vp_token: Vec<Kind<VerifiablePresentation>> = serde_json::from_str(&decoded)?;
-            req.vp_token = Some(vp_token);
-        }
-        if let Some(presentation_submission) = map.get("presentation_submission") {
-            let decoded = urlencoding::decode(presentation_submission)?;
-            let presentation_submission: PresentationSubmission = serde_json::from_str(&decoded)?;
-            req.presentation_submission = Some(presentation_submission);
-        }
-        if let Some(state) = map.get("state") {
-            req.state = Some(state.to_string());
-        }
-        Ok(req)
-    }
-}
-
-/// Authorization Response response object is used to return a `redirect_uri` to
-/// the Wallet following successful processing of the presentation submission.
-#[derive(Debug, Deserialize, Serialize)]
-pub struct ResponseResponse {
-    /// When the redirect parameter is used the Wallet MUST send the User Agent
-    /// to the provided URI. The redirect URI allows the Verifier to
-    /// continue the interaction with the End-User on the device where the
-    /// Wallet resides after the Wallet has sent the Authorization Response.
-    /// It especially enables the Verifier to prevent session fixation
-    /// attacks.
-    ///
-    /// The URI — an absolute URI — is chosen by the Verifier. It MUST include a
-    /// fresh, cryptographically random number to ensure only the receiver
-    /// of the redirect can fetch and process the Authorization Response.
-    /// The number could be added as a path component or a parameter to the
-    /// URL. It is RECOMMENDED to use a cryptographic random value of 128
-    /// bits or more.
-    ///
-    /// # Example
-    ///
-    /// ```http
-    /// redirect_uri": "https://client.example.org/cb#response_code=091535f699ea575c7937fa5f0f454aee"
-    /// ```
-    /// If the response does not contain a parameter, the Wallet is not required
-    /// to perform any further steps.
-    pub redirect_uri: Option<String>,
-
-    /// A cryptographically random number with sufficient entropy used to link
-    /// the Authorization Response to the Authorization Request. The
-    /// `response_code` is returned to the Verifier when the Wallet follows
-    /// the redirect in the `redirect_uri` parameter.
-    pub response_code: Option<String>,
-}
-
-/// Request to retrieve the Verifier's  client metadata.
-#[derive(Clone, Debug, Default, Deserialize)]
-pub struct MetadataRequest {
-    /// The Verifier's Client Identifier for which the configuration is to be
-    /// returned.
-    #[serde(default)]
-    pub client_id: String,
-}
-
-/// Response containing the Verifier's client metadata.
-#[derive(Clone, Debug, Default, Deserialize, Serialize, PartialEq, Eq)]
-pub struct MetadataResponse {
-    /// The Client metadata for the specified Verifier.
-    #[serde(flatten)]
-    pub client: Verifier,
-}
-
-/// Used to define the format and proof types of Verifiable Presentations and
-/// Verifiable Credentials that a Verifier supports.
-///
-/// Deployments can extend the formats supported, provided Issuers, Holders and
-/// Verifiers all understand the new format.
-/// See <https://openid.net/specs/openid-4-verifiable-presentations-1_0.html#alternative_credential_formats>
-#[derive(Clone, Debug, Default, Deserialize, Serialize, PartialEq, Eq)]
-pub struct VpFormat {
-    /// Algorithms supported by the format.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub alg: Option<Vec<String>>,
-
-    /// Proof types supported by the format.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub proof_type: Option<Vec<String>>,
-}
-
-/// OAuth 2 client metadata used for registering clients of the issuance and
-/// wallet authorization servers.
-///
-/// In the case of Issuance, the Wallet is the Client and the Issuer is the
-/// Authorization Server.
-///
-/// In the case of Presentation, the Wallet is the Authorization Server and the
-/// Verifier is the Client.
-#[derive(Clone, Debug, Default, Deserialize, Serialize, PartialEq, Eq)]
-pub struct Verifier {
-    /// OAuth 2.0 Client
-    #[serde(flatten)]
-    pub oauth: OAuthClient,
-
-    /// An object defining the formats and proof types of Verifiable
-    /// Presentations and Verifiable Credentials that a Verifier supports.
-    /// For specific values that can be used.
-    ///
-    /// # Example
-    ///
-    /// ```json
-    /// "jwt_vc_json": {
-    ///     "proof_type": [
-    ///         "JsonWebSignature2020"
-    ///     ]
-    /// }
-    /// ```
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub vp_formats: Option<HashMap<Format, VpFormat>>,
-}
-
-/// The `OpenID4VCI` specification defines commonly used [Credential Format
-/// Profiles] to support.  The profiles define Credential format specific
-/// parameters or claims used to support a particular format.
-///
-/// [Credential Format Profiles]: (https://openid.net/specs/openid-4-verifiable-credential-issuance-1_0.html#name-credential-format-profiles)
-#[derive(Clone, Debug, Deserialize, Serialize, PartialEq, Eq, Hash)]
-pub enum Format {
-    /// W3C Verifiable Credential.
-    #[serde(rename = "jwt_vp_json")]
-    JwtVpJson,
-}
-
-/// OAuth 2.0 Authorization Server metadata.
-/// See RFC 8414 - Authorization Server Metadata
-#[derive(Clone, Debug, Default, Deserialize, Serialize, PartialEq, Eq)]
-pub struct Wallet {
-    /// OAuth 2.0 Server
-    #[serde(flatten)]
-    pub oauth: OAuthServer,
-
-    /// Specifies whether the Wallet supports the transfer of
-    /// `presentation_definition` by reference, with true indicating support.
-    /// If omitted, the default value is true.
-    pub presentation_definition_uri_supported: bool,
-
-    /// A list of key value pairs, where the key identifies a Credential format
-    /// supported by the Wallet.
-    pub vp_formats_supported: Option<HashMap<String, VpFormat>>,
-}
-
-#[cfg(test)]
-mod tests {
-    use insta::assert_yaml_snapshot as assert_snapshot;
-
-    use super::*;
-    use crate::dif_exch::{DescriptorMap, PathNested};
-
-    #[test]
-    fn response_request_form_encode() {
-        let request = ResponseRequest {
-            vp_token: Some(vec![Kind::String("eyJ.etc".into())]),
-            presentation_submission: Some(PresentationSubmission {
-                id: "07b0d07c-f51e-4909-a1ab-d35e2cef20b0".into(),
-                definition_id: "4b93b6aa-2157-4458-80ff-ffcefa3ff3b0".into(),
-                descriptor_map: vec![DescriptorMap {
-                    id: "employment".into(),
-                    format: "jwt_vc_json".into(),
-                    path: "$".into(),
-                    path_nested: PathNested {
-                        format: "jwt_vc_json".into(),
-                        path: "$.verifiableCredential[0]".into(),
-                    },
-                }],
-            }),
-            state: Some("Z2VVKkglOWt-MkNDbX5VN05RRFI4ZkZeT01ZelEzQG8".into()),
-        };
-        let map = request.form_encode().expect("should condense to hashmap");
-        assert_snapshot!("response_request_form_encoded", &map, {
-            "." => insta::sorted_redaction(),
-        });
-        let req = ResponseRequest::form_decode(&map).expect("should expand from hashmap");
-        assert_snapshot!("response_request_form_decoded", &req);
     }
 }
