+++ tests/vci_authorize.rs
@@ -3,9 +3,9 @@
 mod utils;
 
 use base64ct::{Base64UrlUnpadded, Encoding};
+use credibil_vc::oid4vci::provider::StateStore;
 use credibil_vc::oid4vci::state::State;
 use credibil_vc::oid4vci::{self, AuthorizationRequest, AuthorizationResponse};
-use credibil_vc::openid::provider::StateStore;
 use insta::assert_yaml_snapshot as assert_snapshot;
 use rstest::rstest;
 use serde_json::{Value, json};
