+++ tests/vci_create_offer.rs
@@ -3,10 +3,10 @@
 mod utils;
 
 use assert_let_bind::assert_let;
+use credibil_vc::oid4vci::provider::StateStore;
 use credibil_vc::oid4vci::state::{Stage, State};
 use credibil_vc::oid4vci::types::{OfferType, SendType};
 use credibil_vc::oid4vci::{self, CreateOfferRequest, CreateOfferResponse};
-use credibil_vc::openid::provider::StateStore;
 use insta::assert_yaml_snapshot as assert_snapshot;
 use serde_json::json;
 use test_issuer::{CREDENTIAL_ISSUER, NORMAL_USER};
