+++ tests/vci_register.rs
@@ -4,9 +4,9 @@ mod utils;
 
 use chrono::Utc;
 use credibil_vc::oid4vci;
+use credibil_vc::oid4vci::provider::StateStore;
 use credibil_vc::oid4vci::state::{Expire, Stage, State, Token};
 use credibil_vc::oid4vci::types::RegistrationRequest;
-use credibil_vc::openid::provider::StateStore;
 use insta::assert_yaml_snapshot as assert_snapshot;
 use serde_json::json;
 use test_issuer::{CLIENT_ID, CREDENTIAL_ISSUER};
