+++ tests/vp_create_request.rs
@@ -3,10 +3,10 @@
 mod utils;
 
 use assert_let_bind::assert_let;
-use credibil_vc::oid4vp;
+use credibil_vc::oid4vp::provider::StateStore;
 use credibil_vc::oid4vp::state::State;
-use credibil_vc::oid4vp::verifier::CreateRequestRequest;
-use credibil_vc::openid::provider::StateStore;
+use credibil_vc::oid4vp::types::CreateRequestRequest;
+use credibil_vc::oid4vp::{self, CreateRequestResponse};
 use insta::assert_yaml_snapshot as assert_snapshot;
 use serde_json::json;
 
@@ -33,12 +33,12 @@ async fn same_device() {
         "device_flow": "SameDevice"
     });
 
-    let mut request =
+    let mut request: CreateRequestRequest =
         serde_json::from_value::<CreateRequestRequest>(body).expect("should deserialize");
     request.client_id = "http://localhost:8080".into();
 
-    let response =
-        oid4vp::create_request(provider.clone(), &request).await.expect("response is ok");
+    let response: CreateRequestResponse =
+        oid4vp::endpoint::handle("http://localhost:8080", request, &provider).await.expect("ok");
 
     assert_eq!(response.request_uri, None);
     assert_let!(Some(req_obj), &response.request_object);
@@ -86,7 +86,7 @@ async fn cross_device() {
     request.client_id = "http://localhost:8080".into();
 
     let response =
-        oid4vp::create_request(provider.clone(), &request).await.expect("response is ok");
+        oid4vp::endpoint::handle("http://localhost:8080", request, &provider).await.expect("ok");
 
     assert!(response.request_object.is_none());
     assert_let!(Some(req_uri), response.request_uri);
