+++ tests/vp_request_object.rs
@@ -6,11 +6,11 @@ use chrono::Utc;
 use credibil_infosec::jose::jws;
 use credibil_vc::core::Kind;
 use credibil_vc::dif_exch::PresentationDefinition;
+use credibil_vc::oid4vp::provider::StateStore;
 use credibil_vc::oid4vp::state::{Expire, State};
-use credibil_vc::oid4vp::verifier::{
+use credibil_vc::oid4vp::types::{
     ClientIdScheme, RequestObject, RequestObjectRequest, RequestObjectType, ResponseType, Verifier,
 };
-use credibil_vc::openid::provider::StateStore;
 use credibil_vc::{oid4vp, verify_key};
 use insta::assert_yaml_snapshot as assert_snapshot;
 use test_verifier::VERIFIER_ID;
@@ -50,7 +50,7 @@ async fn request_jwt() {
         id: state_key.to_string(),
     };
     let response =
-        oid4vp::request_object(provider.clone(), &request).await.expect("response is valid");
+        oid4vp::endpoint::handle("http://localhost:8080", request, &provider).await.expect("ok");
 
     let RequestObjectType::Jwt(jwt_enc) = &response.request_object else {
         panic!("no JWT found in response");
